requirements.txt
selenium
webdriver-manager
pytest

psycopg2-binary